refactor(coach): extract useCoachChat hook (no behavior change)
Step 3 of COACH_PASS2_PHASE_B_PLAN. Moves the chat lifecycle out of
RideshareCoach.tsx into client/src/hooks/coach/useCoachChat.ts:
SSE streaming, action-tag dispatch (notes refetch + validation banner),
useChatPersistence wiring, useMemory conversation logging, attachments
state + handleFileSelect, controllerRef, messagesEndRef, scrollToBottom.

Component now owns only input field state, audio (until step 4), notes
panel state, and the mic toggle. Two callbacks bridge the hook back to
the component: onStreamComplete (post-stream TTS gating) and onNotesSaved
(notes refetch). fetchNotes and handleStreamComplete are forward-defined
above the useCoachChat call so the closures are available at hook setup.

send() signature widened to send(text, attachmentsOverride?) per plan
literal. Internal variable names (fullResponse, isStreaming) preserved.
The 100ms setTimeout in suggested-question buttons is intentionally kept
to preserve the visual flash of the question populating the input.

Verified: tsc --noEmit clean; gateway /api/health 200; Coach route serves.
Browser smoke test deferred to Melody.
